Refactor RedisConnection (Breaking change!)
Let connection take a socket in a private constructor so it's final and
non-null. This is also the case for _stream.

Refactor the connect methods into static builder methods for creating a new
connection and returning a Command with it.

I removed the RedisParser parser member as it was not used for anything (!)